cloud-backend/auth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
slowapi==0.1.9
